+++ Assets/Scripts/PlayerFSM/FishingMinigameChase_Collision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac8f232668876f845ac7368e8b1bb20e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
