+++ Assets/Scripts/PlayerFSM/FishingMinigameChase_Evasion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c6aedc5ab15685b45bf8c301d1eaf641
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
